diodes behaviour

+++ hardware/main/include/mqtt.hpp
@@ -47,11 +47,11 @@ namespace mqtt {
         }
 
         static uint8_t min_moisture() {
-            return static_cast<uint8_t>(MqttClient::getClient().humidity["validRange"]["min"].get<float>());
+            return static_cast<uint8_t>(MqttClient::getClient().humidity["validRange"]["min"].get<int>());
         }
 
         static uint8_t max_moisture() {
-            return static_cast<uint8_t>(MqttClient::getClient().humidity["validRange"]["max"].get<float>());
+            return static_cast<uint8_t>(MqttClient::getClient().humidity["validRange"]["max"].get<int>());
         }
     };
